#161 Update Renderend setup
