+++ apps/lundin/src/app/modules/renderend/renderend.component.html
@@ -1,3 +1,1 @@
-<div #gameArea class="game-area">
-	<lundin-renderend-game></lundin-renderend-game>
-</div>
+<div #gameHost class="game-host"></div>
