+++ apps/lundin/src/app/modules/renderend/renderend.component.scss
@@ -8,6 +8,6 @@
 	grid-template-rows: calc(100vh - 5rem);
 }
 
-.game-area {
+.game-host {
 	grid-area: game;
 }
