+++ libs/renderend/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./renderend"
 export * from "./renderend-display"
+export * from "./renderend-game"
 export * from "./renderend-input"
 export * from "./state/renderend-action"
